requirements.txt
torch
torchvision
torchaudio
transformers
# diffusers
accelerate
peft
gradio
opencv-python
pillow
numpy
tqdm
scipy
sentencepiece
PyYAML
# diffusers==0.27.2
# huggingface-hub==0.24.5
diffusers==0.26.3
huggingface-hub==0.20.3
datasets



